requirements.txt
langchain-openai
langchain
langgraph
pydantic
python-dotenv
jupyter_client
pyzmq
ipykernel

-e .

#optional
gradio